backend/requirements.txt
flask
flask-cors
boto3
amazon-braket-sdk
waitress
